change2

+++ BBS/WebContent/view.jsp
@@ -87,7 +87,7 @@
 	if(userID != null && userID.equals(bbs.getUserID())){
 %>
 	<a href="update.jsp?bbsID=<%=bbsID %>">수정</a>
-	<a href="deleteAction.jsp?bbsID=<%=bbsID %>">삭제</a>
+	<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID %>">삭제</a>
 <%
 	}
 %>
